Refactor size_t to uint64_t
Refactor printf statements too and now the error checker is being SILLY and not parsing my format strings correctly (I suspect it happens before the macro expansion... ugh)

+++ src/lib/histogram/include/histogram/basic_histogram.h
@@ -1,27 +1,27 @@
 #pragma once
 
 #include <stdbool.h>
-#include <stddef.h>
+#include <stdint.h>
 
 /// @note   I assume no overflow in any of these values!
 struct BasicHistogram {
-    size_t *histogram;
+    uint64_t *histogram;
     /// Maximum allowable length of the histogram
-    size_t length;
+    uint64_t length;
     /// We have seen this before, but we do not track stacks this large
-    size_t false_infinity;
+    uint64_t false_infinity;
     /// We have not seen this before
-    size_t infinity;
-    size_t running_sum;
+    uint64_t infinity;
+    uint64_t running_sum;
 };
 
-const size_t DEFAULT_MAX_HISTOGRAM_LENGTH = 1 << 20;
+const uint64_t DEFAULT_MAX_HISTOGRAM_LENGTH = 1 << 20;
 
 bool
-basic_histogram_init(struct BasicHistogram *me, const size_t length);
+basic_histogram_init(struct BasicHistogram *me, const uint64_t length);
 
 bool
-basic_histogram_insert_finite(struct BasicHistogram *me, const size_t index);
+basic_histogram_insert_finite(struct BasicHistogram *me, const uint64_t index);
 
 bool
 basic_histogram_insert_infinite(struct BasicHistogram *me);
